move policy actions and subject merge functions to common utils folder. Use merge actions and subjects in resourceadm policy editor
