--- frontend/app-development/layout/SettingsModalButton/SettingsModal/utils/tabUtils/policyTabUtils/index.ts
@@ -1,4 +1,0 @@
-export {
-  mergeActionsFromPolicyWithActionOptions,
-  mergeSubjectsFromPolicyWithSubjectOptions,
-} from './policyTabUtils';
